move elements and fix types
